Love Calculator Using Count()
This is one of the funniest, yet toughest codes I have ever attempted to write. This bundle can be used to find the count of the given str in the input values, which later concats and converts to int, to output a score, which then using the IF statements, provides necessary comments. 

IDE - VIsual Studio Code- Jupyter Kernel
Python 3.10
Saved As - JSON